# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gachita API",
        "title": "Gachita API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "로그인",
                "parameters": [
                    {
                        "description": "로그인 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/LoginResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/signup": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "회원가입",
                "parameters": [
                    {
                        "description": "가입 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/UserResponse"
                        }
                    }
                }
            }
        },
...
        "/api/rooms/{id}/matches": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match"
                ],
                "summary": "매칭 목록",
                "parameters": [
                    {
                        "type": "string",
                        "description": "방 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/MatchResponse"
                            }
                        }
                    }
                }
            }
        },
        "/api/rooms/{id}/matches/{matchId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match"
                ],
                "summary": "매칭 상세 (멤버 포함)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "방 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "매칭 ID",
                        "name": "matchId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MatchDetailResponse"
                        }
                    }
                }
            }
        },
...
        "/api/rooms/{id}/stops": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hub_stop"
                ],
                "summary": "거점 목록",
                "parameters": [
                    {
                        "type": "string",
                        "description": "방 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/HubStopResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hub_stop"
                ],
                "summary": "거점 추가",
                "parameters": [
                    {
                        "type": "string",
                        "description": "방 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "거점 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateHubStopRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/HubStopResponse"
                        }
                    }
                }
            }
        },
        "/api/rooms/{id}/stops/{stopId}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hub_stop"
                ],
                "summary": "거점 수정",
                "parameters": [
                    {
                        "type": "string",
                        "description": "방 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "거점 ID",
                        "name": "stopId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "거점 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateHubStopRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/HubStopResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hub_stop"
                ],
                "summary": "거점 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "방 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "거점 ID",
                        "name": "stopId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "CreateHubStopRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "정문"
                },
                "sort_order": {
                    "type": "integer",
                    "x-nullable": true,
                    "example": 0
                }
            }
        },
...
        "HubStopResponse": {
            "type": "object",
            "required": [
                "id",
                "name",
                "room_id",
                "sort_order"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
...
        "LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "LoginResponse": {
            "type": "object",
            "required": [
                "access_token",
                "token_type"
            ],
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
        "MatchDetailResponse": {
            "type": "object",
            "required": [
                "created_at",
                "id",
                "members",
                "room_id",
                "status"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/MatchMemberResponse"
                    }
                },
                "room_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "MatchMemberResponse": {
            "type": "object",
            "required": [
                "from_stop_id",
                "nickname",
                "queue_entry_id",
                "time_end",
                "time_start",
                "to_stop_id",
                "user_id"
            ],
            "properties": {
                "from_stop_id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "queue_entry_id": {
                    "type": "string"
                },
                "time_end": {
                    "type": "string"
                },
                "time_start": {
                    "type": "string"
                },
                "to_stop_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "MatchResponse": {
            "type": "object",
            "required": [
                "created_at",
                "id",
                "room_id",
                "status"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-08-02T09:30:00Z"
                },
                "id": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "MessageResponse": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "SignUpRequest": {
            "type": "object",
            "required": [
                "email",
                "nickname",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "nickname": {
                    "type": "string",
                    "example": "가지"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "UpdateHubStopRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "후문"
                },
                "sort_order": {
                    "type": "integer",
                    "x-nullable": true,
                    "example": 1
                }
            }
        },
        "UserResponse": {
            "type": "object",
            "required": [
                "email",
                "id",
                "nickname"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "nickname": {
                    "type": "string",
                    "example": "가지"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}